+++ CSETWebNg/src/app/app.module.ts
@@ -597,7 +597,7 @@ import { CodeEditorModule, provideCodeEditor } from '@ngstack/code-editor';
 import { ImportComponent } from './import/import.component';
 import { NewAssessmentComponent } from './initial/new-assessment/new-assessment.component';
 import { register as registerSwiper } from 'swiper/element/bundle';
-import { AdminSettingsComponent } from './dialogs/admin-settings/admin-settings.component';
+import { AdminSettingsComponent } from './initial/admin-settings/admin-settings.component';
 import { UserService } from './services/user.service';
 
 registerSwiper();
